requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain-google-genai
langchain-community
google-generativeai
tavily-python
pydantic
fastapi.middleware.cors